Version 3.0.2:     Change:       * Added a new Rule that enables ReMix to Strictly Enforce rules that are usually trusted to be followed by it's connected Users.         * When enabled ReMix will Disconnect Users when they attempt to join a "Locked" party.           * This isn't really a rule, but it's something that can be enforced when Strict Rules is enabled.         * When enabled alongside the "noPet" rule, ReMix will disregard any packets calling a Pet.           * This does not disconnect the User, and may be subject to change.         * Enabling this Rule will allow ReMix to disconnect Users breaking the following rules.           * noPK, allPK, noCheat, noMod, noMigrate, and minV (version).       * Added two more Icons for use within the PlrListWidget.         * This include the Cheater, Modder, and Muted icons.           * The Muted icon is currently not used.       * Renamed the Rules within the Rules Widget UI to be somewhat more descriptive in their function.

+++ src/cmdhandler.cpp
@@ -1469,7 +1469,7 @@ QPair<qint64, TimePeriods> CmdHandler::getTimePeriodFromString(const QString& st
         else //Seconds if no letter is given.
             time = TimePeriods::Seconds;
 
-        duration = static_cast<qint32>( time ) * Helper::strToInt( pStr, static_cast<int>( IntBase::DEC ) );
+        duration = static_cast<qint32>( time ) * Helper::strToInt( pStr, IntBase::DEC );
     }
     return QPair<qint64, TimePeriods>( duration, time );
 }
